+++ packages/core/src/element/edge.ts
@@ -590,18 +590,33 @@ Shape.registerEdge(
   'cubic-vertical',
   {
     curvePosition: [1 / 2, 1 / 2],
+    minCurveOffset: [0, 0],
+    curveOffset: undefined,
     getControlPoints(cfg: EdgeConfig): IPoint[] {
       const { startPoint, endPoint } = cfg;
-      if (cfg.curvePosition !== undefined) this.curvePosition = cfg.curvePosition;
-      if (isNumber(this.curvePosition))
-        this.curvePosition = [this.curvePosition, 1 - this.curvePosition];
+      if (cfg.curvePosition === undefined) cfg.curvePosition = this.curvePosition;
+      if (cfg.curveOffset === undefined) cfg.curveOffset = this.curveOffset;
+      if (cfg.minCurveOffset === undefined) cfg.minCurveOffset = this.minCurveOffset;
+      if (isNumber(cfg.curveOffset)) cfg.curveOffset = [cfg.curveOffset, -cfg.curveOffset];
+      if (isNumber(cfg.minCurveOffset))
+        cfg.minCurveOffset = [cfg.minCurveOffset, -cfg.minCurveOffset];
+      if (isNumber(cfg.curvePosition))
+        cfg.curvePosition = [cfg.curvePosition, 1 - cfg.curvePosition];
+
+      const yDist = Math.abs(endPoint!.y - startPoint!.y);
+      const curveOffset: number[] = cfg.curveOffset
+        ? cfg.curveOffset
+        : Math.abs(yDist) < Math.abs(cfg.minCurveOffset[0])
+        ? cfg.minCurveOffset
+        : [0, 0];
+
       const innerPoint1 = {
         x: startPoint!.x,
-        y: (endPoint!.y - startPoint!.y) * (this as any).curvePosition[0] + startPoint!.y,
+        y: startPoint!.y + yDist * (this as any).curvePosition[0] + curveOffset[0],
       };
       const innerPoint2 = {
         x: endPoint!.x,
-        y: (endPoint!.y - startPoint!.y) * (this as any).curvePosition[1] + startPoint!.y,
+        y: endPoint!.y - yDist * (this as any).curvePosition[1] + curveOffset[1],
       };
       const controlPoints = [innerPoint1, innerPoint2];
       return controlPoints;
@@ -615,17 +630,31 @@ Shape.registerEdge(
   'cubic-horizontal',
   {
     curvePosition: [1 / 2, 1 / 2],
+    minCurveOffset: [0, 0],
+    curveOffset: undefined,
     getControlPoints(cfg: EdgeConfig): IPoint[] {
       const { startPoint, endPoint } = cfg;
-      if (cfg.curvePosition !== undefined) this.curvePosition = cfg.curvePosition;
-      if (isNumber(this.curvePosition))
-        this.curvePosition = [this.curvePosition, 1 - this.curvePosition];
+      if (cfg.curvePosition === undefined) cfg.curvePosition = this.curvePosition;
+      if (cfg.curveOffset === undefined) cfg.curveOffset = this.curveOffset;
+      if (cfg.minCurveOffset === undefined) cfg.minCurveOffset = this.minCurveOffset;
+      if (isNumber(cfg.curveOffset)) cfg.curveOffset = [cfg.curveOffset, -cfg.curveOffset];
+      if (isNumber(cfg.minCurveOffset))
+        cfg.minCurveOffset = [cfg.minCurveOffset, -cfg.minCurveOffset];
+      if (isNumber(cfg.curvePosition))
+        cfg.curvePosition = [cfg.curvePosition, 1 - cfg.curvePosition];
+
+      const xDist = Math.abs(endPoint!.x - startPoint!.x);
+      const curveOffset: number[] = cfg.curveOffset
+        ? cfg.curveOffset
+        : Math.abs(xDist) < Math.abs(cfg.minCurveOffset[0])
+        ? cfg.minCurveOffset
+        : [0, 0];
       const innerPoint1 = {
-        x: (endPoint!.x - startPoint!.x) * (this as any).curvePosition[0] + startPoint!.x,
+        x: startPoint!.x + xDist * (this as any).curvePosition[0] + curveOffset[0],
         y: startPoint!.y,
       };
       const innerPoint2 = {
-        x: (endPoint!.x - startPoint!.x) * (this as any).curvePosition[1] + startPoint!.x,
+        x: endPoint!.x - xDist * (this as any).curvePosition[1] + curveOffset[1],
         y: endPoint!.y,
       };
       const controlPoints = [innerPoint1, innerPoint2];